requirements.txt
Django==3.1.5
django_rest_knox==4.2.0
django_rest_passwordreset==1.3.0
django_rest_swagger==2.2.0
djangorestframework==3.14.0
knox==0.1.14
